requirement.txt
alembic==0.9.9
boto==2.48.0
boto3==1.7.50
botocore==1.10.50
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
click==6.7
cycler==0.10.0
docutils==0.14
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap4==4.0.2
Flask-Migrate==2.1.1
Flask-MySQLdb==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gensim==3.4.0
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.2.2
mysqlclient==1.3.12
nltk==3.3
numpy==1.14.5
pandas==0.23.3
pyparsing==2.2.0
python-dateutil==2.7.3
python-editor==1.0.3
python3-memcached==1.51
pytz==2018.5
requests==2.19.1
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.9.0
six==1.11.0
smart-open==1.6.0
SQLAlchemy==1.2.7
tornado==5.0.2
tqdm==4.23.4
urllib3==1.23
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
